api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.32
alembic==1.13.2
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.4.0
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
httpx==0.27.0
apscheduler==3.10.4
weasyprint==61.2
jinja2==3.1.4

# dev
pytest==8.3.2
pytest-asyncio==0.23.8
ruff==0.5.7
black==24.8.0
isort==5.13.2
